autenticacao+autorizacao JWT completa

+++ api/src/main/java/br/com/digisystem/api/model/Produto.java
@@ -10,6 +10,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
 
+
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,28 +21,31 @@ import lombok.ToString;
 
 @Data
 @Builder
-@EqualsAndHashCode(onlyExplicitlyIncluded = true)
-@AllArgsConstructor 
+@EqualsAndHashCode ( onlyExplicitlyIncluded = true)
+@AllArgsConstructor
 @NoArgsConstructor
 @ToString
 
 @Entity
-public class Produto{
+public class Produto {
 	
+
 	@Id
-	@GeneratedValue (strategy = GenerationType.IDENTITY )
+	@GeneratedValue ( strategy = GenerationType.IDENTITY )
 	@EqualsAndHashCode.Include
 	private int id;
 	
-	@ApiModelProperty(name= "nome", required = true)
+	@ApiModelProperty(name = "nome", required = true)
 	private String nome;
-	private float preco;
+	private double preco;
 	
 	@ManyToMany
-	@JoinTable(
-			name = "produtos_categorias",
-			joinColumns = @JoinColumn (name = "produto_id"),
-			inverseJoinColumns = @JoinColumn(name = "categoria_id")
-			)
+	@JoinTable ( 
+			name="produtos_categorias" ,  
+			joinColumns = @JoinColumn ( name ="produto_id"), 
+			inverseJoinColumns = @JoinColumn( name = "categoria_id" ) 
+		)	
 	private List<Categoria> categorias;
+		
+	//lombok
 }
